routes working, dateFormat

+++ routes/api/userRoutes.js
@@ -1,5 +1,6 @@
 const router = require('express').Router();
-const User = require('../models/User');
+const User = require('../../Models/User');
+const Thought = require('../../Models/Thought');
 
 // GET all users
 router.get('/', async (req, res) => {
@@ -20,16 +21,24 @@ router.post('/', async (req, res) => {
     res.json(user);
 });
 
-// PUT update user by id
+// PUT update user's username by id
 router.put('/:id', async (req, res) => {
-    const user = await User.findByIdAndUpdate(req.params.id, req.body, { new: true });
+    const { username } = req.body;
+    const user = await User.findByIdAndUpdate(req.params.id, { username }, { new: true });
     res.json(user);
 });
 
 // DELETE user by id
 router.delete('/:id', async (req, res) => {
-    const user = await User.findByIdAndDelete(req.params.id);
-    res.json(user);
+    const user = await User.findById(req.params.id);
+    if (!user) {
+        return res.status(404).json({ message: 'No user found with this id!' });
+    }
+    // Remove user's thoughts
+    await Thought.deleteMany({ username: user.username });
+    // Remove user
+    await User.findByIdAndDelete(req.params.id);
+    res.json({ message: 'User and associated thoughts deleted!' });
 });
 
 // POST add friend to user's friend list
